[BUGFIX beta] Only add deprecated container after create when present.
When the object returned from calling `.create` on a non-extendable
factory does not contain a `container` property, do not set it.

Previously, in Ember 2.2.0 it was possible to use non-extendable
factories that were frozen after creation. After the changes to add a
mock container to `.create` and swap that with the deprecated
property after `.create` was finished, we trigger errors for anything
that is frozen or sealed.

(cherry picked from commit b6e783d6fb4310d020feacb12052cbcac35ec7f8)