Default autoboot to false when in node environment
The code paths inside Application.init, within the ember-application-visit flag
are never hit by any of the tests. Furthermore, autoboot must be explicitly set
to false when running in node to allow passing in a url when booting an app
instance.

Thus, make it default to false when in node, true when in the browser, to
abstract this concern away from the user.