Use EmptyObject for `dictionary(null)`.
A huge number of usages of `ember-metal/dictionary` are using `null` as
the parent. Creating a `new EmptyObject()` is quite a bit faster than
`Object.create(null)` (see comments in `ember-metal/empty_object` for
details).

Note: we are working around an issue with QUnit due to using constructor
value of `undefined` (they support `null` as of 1.20). Perhaps we should
tweak to use `null`, but that seems out of scope for this change.