Do not override `OWNER` if already present on child view.
Previously, `linkChild` would always set the new child view's `OWNER` to
its own `OWNER`.  Unfortunately, when dealing with Engines mounted at a
given outlet this means that any components and whatnot rendered in that
engines template have their `OWNER` changed after init (just before
render).

The intent of the original code was to ensure that all child views had
an `OWNER` (actually it was `.container` at the time). The
`ViewNodeManager` and `ComponentNodeManager`'s now handle properly
creating the views/components with the right `OWNER` so this code is
mostly no longer needed (except for edge case legacy support). The guard
added here ensures that we are only setting `OWNER` if it wasn't
previously set (which is that one legacy edge case).