[BUGFIX canary] Support tagless components in fastboot
Attempting to render a component with `tagName: ''` in fastboot kills the fastboot server with a stack trace that looks like:

```
undefined:55214
      _emberViewsSystemJquery.default(rootElement).off('.ember', '**').removeC
                                     ^
TypeError: undefined is not a function
    at exports.default._emberRuntimeSystemObject.default.extend.destroy (<anonymous>:55214:38)
    at superWrapper [as destroy] (<anonymous>:33934:22)
    at <anonymous>:11887:16
    at eachDestroyable (<anonymous>:12166:9)
    at Object.Container.destroy (<anonymous>:11885:7)
    at Object.Backburner.run (<anonymous>:11256:25)
    at Object.run [as default] (<anonymous>:31959:27)
    at exports.default._emberMetalMixin.Mixin.create.willDestroy (<anonymous>:44068:36)
    at superWrapper (<anonymous>:33934:22)
    at _emberRuntimeSystemObject.default.extend.willDestroy (<anonymous>:15925:19)

```

The root cause is that initializing a tagless component causes a lookup of `event_dispatcher:main`. This actually succeeds, and the `EventDispatcher` doesn't try to do anything, but then during container teardown the EventDispatcher's `destroy` method throws the above exception.

To make any future bugs in this class more obvious, I added an assertion in `EventDispatcher` so that it will fail faster if you try to instantiate it in fastboot mode.