[BUGFIX release] ArrayProxy should not be a computed property
Fixes #12475.

This commit restores the contract of ArrayProxy content being a simple
property instead of a computed property. This prevents accidental
clobberings such as in ember-data:

https://github.com/emberjs/data/blob/v2.3.3/addon/-private/system/record-arrays/record-array.js#L44

This commit also removes a private hook `arrangedContentArrayWillChange`.
In the interest of keeping this bugfix commit minimal, we will continue
removing the will* hooks on canary.