Implement Ember semantics for {{#if}}/{{#unless}}
Also ported most of the basic test cases to the new format. I expect the
super classes here can be extracted into a different file and we should
be able to reuse all the test cases for the inline (if …) and (unless …)
helpers as well, because of the test cases go through the `templateFor`
abstraction.

I have not figured out a good way to abstract the “without inverse”
tests, so they are duplicated in the leaf subclasses.

The rest of the tests in the `if_unless_test` requires some more careful
reading/git-history-digging to see if they are still relevant, which I
do not have time for at the moment (maybe someone else can pick up where
I left off).

In any case, implementing the Ember semantics should be enough to
unblock porting other tests that depends on them. I did not implement
the inline `(if …)` and `(unless …)` helpers or port their tests, but
that should be pretty easy given the `{{concat}}` example and the
`toBool` function in here.

Note: please don’t add the inline `(if …)` and `(unless …)` tests here.
Instead, please add them to `integration/helpers/if-unless-test.js` (and
extract the shared test cases to a different file too).