Better component layout builder API for Ember
This commit continues the process of exposing a minimal public API for
building component layouts outside of Glimmer.

We need to support using the layout as-is (“Glimmer components”) as well
as assembling the layout from dynamic information like tagName,
classNames, attributeBindings, etc.

Even for Glimmer components, we need a way to express adding additional
attributes (like id=“ember123” and class=“ember-view”).

The goal is to avoid leaking the statement syntax API or the appending
opcode API, which until now were necessary to construct a dynamic
layout.