[BUGFIX release] Fix legacy addon deprecations
The signature of `Ember.deprecate` is `message`, `test`, `options`,
where `test` is `If falsy, the deprecation will be displayed.` Here,
we did the opposite, causing the deprecation message to be shown only
when the addons are NOT installed.

This commit fixes the issue, and also moved the deprecation to first-
boot. This gives users (and our test harness) a chance to register
custom handlers before we issue the warning.

Also added tests.
(cherry picked from commit 59e9045d9784dcdafe1cbd164f042528848626dd)