[BUGFIX release] Fix mergedProperties at create time modifying proto
When `mergedProperties` are passed into an object at `.create()` time,
the objects prototype is inadvertently modified, which is then
reflected on all instances of the object. Simply passing in a new
object during `assign()` resolves this problem (per krisselden). Also
added a test for this use case, where `objA` results in 3 when the
proto is modified. Fixes #11714