Rearchitect #each implementation
Previously, the strategy for acquiring a key and its associated
updatable reference for a given iteration item was hardcoded in
glimmer-reference.

This is unacceptable, as the reference type is intended to provide
object-model-specific semantics such as those provided by Ember, but
also libraries like immutable.js.

This change decouples these strategies from the iterator implementation,
allowing the environment to participate.

The environment must now implement an iterableFor hook, which produces
an iterable for a given array-like object and #each arguments. This
iterable produces an iterator per render (and revalidation). The
iterator yields { key, value }. The iterable also encodes the policy
for producing and updating a reference for a given value.