Implement Ember semantics for Glimmer {{#each}}
As usual, we only ported enough tests to test the system is working, and
we need to finish porting the rest of the tests (we actually haven’t
looked into the existing each tests yet – we just borrowed the {{#with}}
tests as a starting point). However, since we are now sharing the same
basic test cases from the conditionals test, we already covered a lot of
ground from-the-get-go.