Glimmer dynamic components (component helpers)
The tests are copied from the (new) curly component tests. The first
test case has been split to test the dynamic component switching, but
this is otherwise just a straight port of the existing tests (which is
itself not the complete set of components tests we have).

As usual, we need to come back and port the rest of the tests.

The foundation for this feature is exhaustively matrix-tested in
Glimmer, including the component switch scenario.