Refactor dynamic scope
This change allows the environment to supply a custom dynamic scope
class rather than trying to model dynamic scope as an array with a fixed
set of slots. Unlike the static scope, which has an arbitrary number of
slots based on the program, dynamic scope has a fixed shape throughout
the program.