Don’t insert unnecessary empty comment for lists
Since 005ba6e, we are handling empty lists the same way we handle
regular blocks (`{{#if}}`, etc), using the `EnterOpcode` infrastructure.

However, we still have an empty comment at the end of every (non-empty).
This is only used as a marker for moves/insertions (at the end of the
list) during updates.

We can remove this requirement by inserting and removing the empty
marker comment before and after the update step. This removed the minor
incompatibility with HTMLBars.