Improve dynamic scopes
This allows us to just pass the initial dynamic scope in the VM init
process, instead of using a callback. We also changed the component
builder API so that you can only say whether you require a new dynamic
scope, but do the actual binding inside ComponentManager#create (and
friends).