Fix loc info for content statements with leading newlines.
Given the following:

```
<div>
  {{#if isSingleError}}
    Single error here!
  {{else if errors}}
    Multiple errors here!
  {{else}}
    No errors found!
  {{/if}}
</div>
```

The `ElementNode` for the `<div>` was indicating that its
`loc.end.column` was 12. This is ultimately due to how Handlebars
handles trailing newlines after the `{{/if}}`. We were properly detecting
that the node was `rightStripped` and adding the number of stripped
newlines to the `tokenizer.line` number, but we did not reset the
`tokenizer.column` also.

The fix is to utilize the `ContentStatement`'s `loc` information to
reset the tokenizer's current position.