Adding Performance.md
I think it would be a good idea to document the known perf bottlenecks.
This will allow maintainers to point to perf guidelines in PRs instead
of relying on tribal knowledge. Luckily the angular folks had [something
like this](https://github.com/angular/angular/wiki/performance) already so I took advantage of that.

This serves as a rough draft and I would like feedback on it from
@wycats, @stefanpenner, and @krisselden.