Kill packages/node_modules
Previously, there was no way to use the “node” module resolution strategy
without using the `node_modules` directory.

This meant that we were forced to store all packages in `node_modules`,
in order to get TypeScript and VSCode to resolve modules correctly.

Unfortunately, this had several unfortunate side effects (in addition to
the obvious unnecessary and confusing directory): GitHub automatically
suppresses `node_modules` changes in diffs and does not index into that
directory (meaning GitHub code search does not work).

In short, this use of `node_modules`, while it worked effectively to
convince TypeScript and VSCode of the correct semantics, abused and
overloaded the conventional meaning of the directory, with predictable
effects on GitHub.

Recently, the TypeScript team added the ability to express “path
mappings”, which gave us the ability to map the node module resolution
strategy into a new directory, which we have done here.

Problem solved. Thanks TypeScript!