[Glimmer2] Migrating log helper
This commit migrates the log helper to Glimmer2 and moves the tests over as well.
Tests located in debug_test.js in ember-htmlbars have been removed, as they are duplicates.

There is a 'FIXME' in the log helper that will need to be fixed once
basic component rendering has been addressed.