Implement `ConstReference` optimization
A const reference cannot change, so we do not need to emit an updating
opcode for them.

For example:

```has
{{#if true}}...{{/if}}
```

In this case, the conditional is a literal, so it is impossible for the
the {{#if}} to flip from default to inverse on the update pass, hence
there is no need to emit an `Assert` for the conditional value.

While this in itself is pretty rare, it is quite common for this to
happen indirectly. For example:

```hbs
{{!-- link-to component layout --}}
<a href="{{@url}}" title={{@title}}>{{@text}}</a>
```

```hbs
{{!-- link-to usage --}}
{{link-to url="http://static.string" title="does not change" text="also does not change"}}
```

In this case, there is no need to emit updating opcodes for the
attributes and content curlies.

This optimization also composes nicely, allowing things like `EmberID`
or `hasBlock` to return const references and take advantage of the same
optimization to drop unnecessary update work.