[BUGFIX release] Add mut to attributes in contextual components
As explained in issue #13061, given a component `change-button` with template:

```hbs
<button {{action (action (mut val) 10)}}>Change to 10</button>
```

If it is invoked as in the following template:

```hbs
{{change-button val=value}}
<span class="value">
  {{value}}
</span>
```

Clicking on the button will cause `.value` to contain `10`.

On the other hand, invoking it via a closure component:

```hbs
{{component (component "change-button" val=value)}}
<span class="value">
  {{value}}
</span>
```

Clicking on the button will not change the text in `.value`.

This PR adds a transform that adds `mut` to params and hash values.

(cherry picked from commit 31bb7a2eb746ed677d568cfd4aa5a79774cc4ef7)