Remove `latest` from `chrome` sauce labs tests.
This is erroring for all PR's saying that `latest` is not a valid
version. We need to dig in and/or report an issue to the Sauce Labs
folks, but this should gets us back to a passing build.

(cherry picked from commit a194468600a0fbd511e13c9b7c38b46d2cd9bb25)