Cleanup and introduce tag-based dirty-tracking
Cleanup:

- Remove `destroy` and `isDirty` from `Reference`
- Split `glimmer-reference` into `glimmer-reference` and `glimmer-object-reference`

The tagging system allows parts of the system to opt-into a more
fine-grained dirty-tracking, replacing the previous `isDirty` system
(which never actually did anything).

In the new system, if a consumer of a reference is interested in cache
its value, they can capture it along with the current "revision" of the
reference, which they can later use to revalidate the freshness of the
cached value.