[BUGFIX release] Ensure normal methods are `_super` wrapped.
Prior to this, if a method was not included in the superclasses
prototype, we would skip wrapping it with its own `_super`. However,
that means that if the method is called from within another `_super`
wrapped method it would not properly reset the `_super` causing
incorrect results.

This ensures that all methods are super wrapped if they contain
`_super` (and `Function.prototype.toString` is present).