[DEPRECATE release] Deprecate usage of `{{render}}` with a block.
This was not intended to be supported and is rarely used. Adding support
for this in tildeio/glimmer will be pretty hard (due to very bizarre
`controller` and target values for the block).

This deprecation is intended to help guide anyone actually using
this functionality before upgrading to 2.5.0 or later.