[BUGFIX canary] Remove Glimmer dependency in `ember-htmlbars`
Since 5efb731, we have removed the `{{concat}}` implementation in
`ember-htmlbars` and moved it into `ember-glimmer` with a symlink.

In c655638, we introduced a dependency on Glimmer (`normalizeTextValue`)
in the new implementation, preventing the helper from working when
the feature flag is off.

This essentially reverts both of those changes and brings back the old
implementation for `ember-htmlbars`.