Support for htmlSafe/SafeString
This creates a new type of Append opcode that handles the value when
it is a SafeString and also when it is a normal string. For SafeString
it enters into "html" mode, and if the input changes into a normal
string it switches into "text" mode. It can switch freely between the
two modes and handles updates accordingly.