Fix toString to not add the property on the class
When calling `toString()` method on an Ember Object, it attaches the `toString` as a class property for the first time. This does not seem right since this should be on the prototype of the class. We could memoize the value on the class meta data.

In v2.5.0 and above, when calling `.set()` on an Ember Object, you end up also attaching the `toString()` property on the class. This [PR](https://github.com/emberjs/ember.js/commit/3a2486eb911ad77d2ab1a4324c1aacef08c156da) calls `toString()` on the object during setting any property. Therefore, when you walk through the object properties you end up getting an additional property on the object.

cc: @krisselden @chadhietala