[BUGFIX canary] Ensure `Ember.ENV` is present.
This adds a non-enumerable property to `Ember` for `ENV`.
Accessing/updating `ENV` after initial boot is not a supported
operation, however ember-cli-htmlbars was using this private API to
setup the various flags (i.e. `EXTEND_PROTOTYPES` and friends).

An update will come shortly for `ember-cli-htmlbars`, but this PR
prevents updating to canary Ember from being a breaking change for users
of current versions of ember-cli-htmlbars.