[BUGFIX beta] Fix subexpr `{{if}}` / `{{unless}}` to use protocol.
Prior to this change, the inline `{{if` / `{{unless` helpers would not
recalulate at the same time as the non-inline versions (because they
were not respecting the `isTruthy` and `.length` checks properly).

This fixes that by properly wrapping the generated stream with those
dependencies (sharing the same code as the block versions of `{{if` /
`{{unless`).

The tests for this were already written, but disabled when running under
HTMLBars. This enables them for running properly in both HTMLBars and
Glimmer engines.

(cherry picked from commit dd3893704b04185d69eee192078b0629881a4b8c)