Implement text node merging fix w/o moving nodes
The initial implementation of the text node merging fix was to
`innerHTML` content then move the nodes of that temporary parent to the
correct location. This changes the implementation to instead inject a
placeholder comment (if necessary) after a previous text node, then
remove that injected comment after the base implementation runs.

Only the previous text node must be protected from merging, since the
current base implementation of `insertHTMLBefore` already protects from
merging with a nextSibling text node.