[BUGFIX beta] Produce a nice message when undefined is passed to closure action helper
If the action helper is passed an undefined value from `attrs` (as in `(action
attrs.some-undefined-value)`) the output message was a bit cryptic. This
creates a nicer error message for the user.

Take into account that the main problem is `attrs` containing the values and
not Streams, therefore the error cannot show a good label.