Better dynamic component infrastructure
When transforming syntax into a dynamic component invocation, it is
often the case that you also need to control over what the args the
component should be invoked with. For example, in Ember:

```
{{component dynamicName some extra="..." args=foo}}
```

Here, the first positional arg (`dynamicName`) is only used internally
to lookup the component definition and should not be seen by the
component in its invocation args. This change allows the host to
separate the args for the component lookup and the component invocation.