Split registrations between Application and Engine classes.
Registrations that have proven to be necessary for all engines,
based on work in the ember-engines addon, are moved from the
Application class to the Engine class.

This eliminates the need for overrides in ember-engines, while tests
still ensure that applications are initialized with the same entries in
their registries as always.