Properly handle trailing whitespace controls.
Previously, we were only handling the Handlebars default "eating" of a
newline after a block statement but did not properly handle `~}}` at the
end of a block statement.

This updates to also calculate the column offsets (and adds tests).

Given the following:

```hbs
{{#if foo~}}
  <div></div>
{{/if}}
```

Before these changes, the `<div>` would have a starting column of `0`.

After these changes, the `<div>` has the correct starting column of `2`.