Add lazy evaluation tests
In both `{{#if truthy}}...{{else}}{{some-helper}}{{/if}}` and `{{if
truthy ... (some-helper)}}`, `some-helper` should NOT be evaluated.
This is especially important when you are trying to use the `if`
helper to guard for `null`s etc.

They already works in Glimmer. The block form works in HTMLBars, but
not the inline form.

See #13589