[BUGFIX release] Fix inlining of superWrapper.
This fixes the inlining of superWrapper.  The bug that copying the arguments was addressing has been fixed for long enough that more users are affected by fix than the fix benefits.  This also fixes a deopt in the megamorphic load of toString, since we store metadata on functions we have many shapes of functions.