[Glimmer 2] Remove tests related to legacy ember-test-helpers
These tests are specifically testing patterns that used to be used in
ember-test-helpers. Since these tests have been introduced
ember-test-helpers has moved to the container pattern with
`lookupFactory`. In general we need to start migrating tests away from
things that did not come out of the container with injections on them.

This is because we need to remove this hardcoded `InteractiveRenderer`
[here](https://github.com/emberjs/ember.js/blob/master/packages/ember-views/lib/views/core_view.js#L46-L52)