[Glimmer 2] Remove invalid tests for {{link-to}} without router.
These tests are not properly emulating the behaviors of
ember-test-helpers. In these tests, the application registry has not
been properly setup/created since it is done in `_bootSync` but the test
harness defers readiness and never truly boots the application (causing
the registry to not be setup).

The specific scenario of `{{link-to}}` working without a router
available is properly tested
[here](https://github.com/emberjs/ember.js/blob/81102a1f49b2b607ddf290ad17b2d1c067c85b6c/packages/ember-glimmer/tests/integration/components/link-to-test.js#L22-L28).