Refactor attribute setting
This refactors attributes and creates 2 distinct types: Attribute and
AttributeChangeLists. The Attribute object is responsible for knowing when
the attribute needs to be set and updated. The AttributeChangeList is the
type that actual does the setting/updating. This distinction allows for
user land code to provide their own policies.

This also includes tests for style warnings and also escaping hrefs.