`assert` does not take a function
In a few places, we are incorrectly passing a function to `assert`.
Since the `assert` function does not actually take a function, it just
treats it as a truthy input, hence these asserts never actually fired.
This error is mostly masked by the fact that we are also using the wrong
Component class in some of the Glimmer tests.

This also removes the aria-role from the default `attributeBindings` for
curly components in Glimmer. Instead, it is implemented in a lower-level
hook, similar to how `elementId` works. This allows us to avoid parsing
the `attributeBindings` micro-syntax in most components and should make
component creation a tiny bit faster.