[FEATURE ember-metal-weakmap] Add a WeakMap polyfill
This patch exposes the Ember's private WeakMap polyfill under a feature
flag. This implementation has a small but important caveat. It assumes
that the weak map will live longer (in the sense of garbage collection)
than all of its keys, otherwise it is possible to leak the values stored
in the weak map.

This patch also includes a few improvements:

- implement WeakMap constructor
- implement WeakMap toString
- guard against non-Object keys