[glimmer2] fix unbound helper test
Before, the test rendered the following:

```handlebars
<div data-foo="{{if cond1 \"T1\" \"F1\"}}"></div>
<div data-foo="{{if cond2 \"T2\" \"F2\"}}"></div>
```

The test was only asserting against the value of the `data-foo` of the
first div, causing it fo fail because it was missing the value of the
`data-foo` in the second div.

This is because the test was using `wrappedTemplateFor`, which
is a convenience method which only expects *one* template to be
rendered at a time.

Instead, we use `wrapperFor` and two `templateFor` calls to instead
create only *one* div and assert against it. What's rendered in the test
now looks like:

```handlebars
<div data-foo="{{if cond1 \"T1\" \"F1\"}}{{if cond2 \"T2\" \"F2\"}}"></div>
```

refs #13644