Remove usage of `targetObject` computed property.
This PR removes the `targetObject` CP in favor of simply looking up the
correct `targetObject` when needed. This simplifies the code path for
`sendAction` and removes some overloading of the term "controller"
(before this change it could mean an actual outlet's controller or the
parentView.controller which may be a component and not a "real"
controller).