Fix outlet states
We forgot to cache the `lastState`, so everything was invalidating all
the time. However, we were supposed to look at `state.render.{controller,template}`,
but we did `state.{controller,template}` instead (which are, of course,
always `undefined`). This actually made it "stable" some of the time,
because the check was `oldStuff === newStuff`. So, as a whole, it more
or less behaves randomly 😗

Also, since Glimmer now accepts `null`s as "no component definition", we
can simplify the code a bit here.