[glimmer2] fix each key collision detection
When using `{{#each things key="key" as |thing|}}`, glimmer2 integration
did not update the DOM correctly when `things` was an array with objects
that shared the same `key` value. For example:

```javascript
this.set('data', [
  {
    name: 'Yehuda' }
  },
  {
    name: 'Jenn'
  },
  {
    name: 'Yehuda'
  }
]);
```

and the following template:

```handlebars
{{#each data key="name" as |person|}}
  {{person.name}}
{{/each}}
```

In the above example, we want each item in the list to render, even though two objects have the same
value for the key (Notice there are two "Yehuda"'s above.

As the iterator goes through the array, it will cache value of the `key` `name`.
In order to make sure the cache updates correctly, we assign it a unique key.