Propagate currly trustworthiness
We need to know the calling side of the ref e.g. either double or triple
curlies.

To keep the wire-format as small as possible we actually do want a
TrustingAttr type. This type is just delegate to DynamicAttr, but slots
the isTrusting field as "true". This is a tradeoff that I think makes
sense as we pay the cost of an additional type, but we don't bloat the
wire-format with bools for each attribute's trustworthiness.