clear Bounds when destroying RenderResult's
In Glimmer, elements are not implicitly removed from the DOM when their
destroy hooks are called. Instead of removing each DOM node individually
which can be very ineffecient, Ember's integration with Glimmer uses
higher levels for disconnecting large number of nodes.

Two examples are the OutletView and then the `RenderResult` which lives
inside Glimmer itself. When `RenderResult` and `OutletView` are called,
they should remove themselves from the DOM, thus removing all their
children DOM nodes.

The `Bounds` object on `RenderResult` was not being cleared when the
`RenderResult` was cleared, meaning that when you ran
`Ember.ApplicationInstance.destroy()` in an Ember test, DOM nodes would
linger in the fixture element for tests.

In Ember CLI, the generated blueprint for acceptance tests is to run
`Ember.Application.destroy()` so that no two tests share the same app
instance. Indeed, Ember's tests also use this. Leaving lingering DOM
nodes were causing several failures in Ember's integration with
Glimmer2.

Thanks @chancancode for helping me understand more of the architecture
of Glimmer2 and debugging this with me! :tada:

refs emberjs/13807