Move `template` method to respective engine packages.
This is currently exposed as `Ember.HTMLBars.template` and
`Ember.Handlebars.template` (we should probably figure out a more
generic place for this in both globals and modules land), and is the
only piece of the `ember-template-compiler` that is actually needed at
runtime when the `dist/ember-template-compiler.js` file is not present.

This paves the way to removing `ember-template-compiler` packge from
`ember.debug.js` and `ember.prod.js` (without changing any public API's
or expectations).